dirty but works
