+++ logs.js
